+++ src/apps/content-editor/src/app/views/ItemEdit/Analytics/views/SinglePageAnalyticsView.tsx
@@ -33,102 +33,25 @@ import {
   convertSecondsToMinutesAndSeconds,
   findValuesForDimensions,
   generateReportRequests,
+  getDateRangeAndLabelsFromParams,
 } from "../utils";
 import { Metric } from "../components/Metric";
 import { useGetAnalyticsPropertiesQuery } from "../../../../../../../../shell/services/cloudFunctions";
-import { PropertiesDialog } from "../components/PropertiesDialog";
+import { PropertiesDialog } from "../../../../components/PropertiesDialog";
 import instanceZUID from "../../../../../../../../utility/instanceZUID";
 import { NotFound } from "../../../../../../../../shell/components/NotFound";
 import SettingsIcon from "@mui/icons-material/Settings";
 import WarningRoundedIcon from "@mui/icons-material/WarningRounded";
 import CompareArrowsRoundedIcon from "@mui/icons-material/CompareArrowsRounded";
 import { CompareDialog } from "../components/CompareDialog";
-
-const getDateRangeAndLabelsFromPreset = (
-  preset: DateFilterValue
-): [Moment, Moment, string, string] => {
-  if (preset.type === "daterange") {
-    const value = preset.value as DateRangeFilterValue;
-    return [
-      moment(value.from, "YYYY-MM-DD"),
-      moment(value.to, "YYYY-MM-DD"),
-      moment(value.from).format("ddd D MMM"),
-      moment(value.to).format("ddd D MMM"),
-    ];
-  } else {
-    switch (preset.value) {
-      case "today":
-        return [moment(), moment(), "Today", "Yesterday"];
-      case "yesterday":
-        return [
-          moment().subtract(1, "days"),
-          moment().subtract(1, "days"),
-          "Yesterday",
-          "Day Before Yesterday",
-        ];
-      case "last_7_days":
-        return [
-          moment().subtract(7, "days"),
-          moment().subtract(1, "days"),
-          "Last 7 Days",
-          "Prior 7 Days",
-        ];
-      case "last_14_days":
-        return [
-          moment().subtract(14, "days"),
-          moment().subtract(1, "days"),
-          "Last 14 Days",
-          "Prior 14 Days",
-        ];
-      case "last_30_days":
-        return [
-          moment().subtract(30, "days"),
-          moment().subtract(1, "days"),
-          "Last 30 Days",
-          "Prior 30 Days",
-        ];
-      case "last_3_months":
-        return [
-          moment().subtract(90, "days"),
-          moment().subtract(1, "days"),
-          "Last 3 Months",
-          "Prior 3 Months",
-        ];
-      case "last_12_months":
-        return [
-          moment().subtract(365, "days"),
-          moment().subtract(1, "days"),
-          "Last 12 Months",
-          "Prior 12 Months",
-        ];
-      case "this_week":
-        return [moment().startOf("week"), moment(), "This Week", "Last Week"];
-      case "this_year":
-        return [moment().startOf("year"), moment(), "This Year", "Last Year"];
-      case "quarter_to_date":
-        return [
-          moment().startOf("quarter"),
-          moment(),
-          "This Quarter",
-          "Last Quarter",
-        ];
-      default:
-        return [
-          moment().subtract(14, "days"),
-          moment().subtract(1, "days"),
-          "Last 14 Days",
-          "Prior 14 Days",
-        ];
-    }
-  }
-};
+import { AnalyticsDateFilter } from "../../../../components/AnalyticsDateFilter";
+import { AnalyticsPropertySelector } from "../../../../components/AnalyticsPropertySelector";
 
 type Props = {
   item: any;
 };
 
 export const SinglePageAnalyticsView = ({ item }: Props) => {
-  const [showPropertiesDialog, setShowPropertiesDialog] = useState(false);
   const [showCompareDialog, setShowCompareDialog] = useState(false);
   const [params, setParams] = useQueryParams();
   const compareItemZUID = params.get("compare");
@@ -152,38 +75,8 @@ export const SinglePageAnalyticsView = ({ item }: Props) => {
     (property: any) => property.name === propertyId
   );
 
-  const activeDateFilter: DateFilterValue = useMemo(() => {
-    const isPreset = Boolean(params.get("datePreset"));
-    const isDateRange =
-      Boolean(params.get("to")) &&
-      Boolean(params.get("from")) &&
-      params.get("to") !== params.get("from");
-
-    if (isPreset) {
-      return {
-        type: "preset",
-        value: params.get("datePreset"),
-      };
-    }
-
-    if (isDateRange) {
-      return {
-        type: "daterange",
-        value: {
-          from: params.get("from"),
-          to: params.get("to"),
-        },
-      };
-    }
-
-    return {
-      type: "",
-      value: "",
-    };
-  }, [params]);
-
   const [startDate, endDate, dateRange0Label, dateRange1Label] =
-    getDateRangeAndLabelsFromPreset(activeDateFilter);
+    getDateRangeAndLabelsFromParams(params);
 
   const {
     data: ga4Data,
@@ -231,40 +124,6 @@ export const SinglePageAnalyticsView = ({ item }: Props) => {
     start_date: "05/25/2023",
     affectedZUID: itemZUID,
   });
-
-  useEffect(() => {
-    if (!params.get("datePreset") && !params.get("from") && !params.get("to")) {
-      handleDateFilterChanged({ type: "preset", value: "last_14_days" });
-    }
-  }, []);
-
-  const handleDateFilterChanged = (dateFilter: DateFilterValue) => {
-    switch (dateFilter.type) {
-      case "daterange": {
-        const value = dateFilter.value as DateRangeFilterValue;
-
-        setParams(value.to, "to");
-        setParams(value.from, "from");
-        setParams(null, "datePreset");
-        return;
-      }
-      case "preset": {
-        const value = dateFilter.value as string;
-
-        setParams(value, "datePreset");
-        setParams(null, "to");
-        setParams(null, "from");
-        return;
-      }
-
-      default: {
-        setParams(null, "to");
-        setParams(null, "from");
-        setParams(null, "datePreset");
-        return;
-      }
-    }
-  };
 
   const isLoading = isFetching || instanceSettingsFetching || isFetchingCompare;
 
@@ -332,31 +191,7 @@ export const SinglePageAnalyticsView = ({ item }: Props) => {
         </Box>
         <Box display="flex" justifyContent="space-between">
           <Box display="flex" gap={1.5}>
-            {isLoading ? (
-              <Skeleton variant="rectangular" width="137px" height="28px" />
-            ) : (
-              <DateFilter
-                clearable={false}
-                value={activeDateFilter}
-                onChange={handleDateFilterChanged}
-                hideCustomDates
-                withDateRange
-                extraPresets={[
-                  {
-                    text: "This Week (Sun - Today)",
-                    value: "this_week",
-                  },
-                  {
-                    text: "This Year (Jan - Today)",
-                    value: "this_year",
-                  },
-                  {
-                    text: "Quarter to Date",
-                    value: "quarter_to_date",
-                  },
-                ]}
-              />
-            )}
+            <AnalyticsDateFilter showSkeleton={isLoading} />
             {isLoading ? (
               <Skeleton variant="rectangular" width="137px" height="28px" />
             ) : compareItemZUID ? (
@@ -391,49 +226,7 @@ export const SinglePageAnalyticsView = ({ item }: Props) => {
               </Button>
             )}
           </Box>
-          <Box display="flex" gap={0.5} alignItems="center">
-            {isLoading ? (
-              <Skeleton variant="rectangular" width="410px" height="8px" />
-            ) : (
-              <Link
-                href={`${propertyData?.dataStreams?.[0]?.webStreamData?.defaultUri}${item.web.path}`}
-                target="__blank"
-                sx={{
-                  maxWidth: "440px",
-                  direction: "rtl",
-                  whiteSpace: "nowrap",
-                  overflow: "hidden",
-                  textOverflow: "ellipsis",
-                  display: "block",
-                }}
-              >
-                {`${
-                  propertyData?.dataStreams?.[0]?.webStreamData?.defaultUri
-                }${item.web.path?.slice(0, -1)}`}
-              </Link>
-            )}
-            {isLoading ? (
-              <Skeleton variant="rectangular" width="38px" height="22px" />
-            ) : (
-              <Button
-                data-cy="analytics-settings"
-                onClick={() => setShowPropertiesDialog(true)}
-                size="small"
-                variant="outlined"
-                color="inherit"
-                sx={{
-                  height: "22px",
-                  width: "38px",
-                  minWidth: "unset",
-                }}
-              >
-                <SettingsIcon
-                  color="action"
-                  sx={{ width: "18px", height: "18px" }}
-                />
-              </Button>
-            )}
-          </Box>
+          <AnalyticsPropertySelector path={item.web.path} />
         </Box>
         <Box
           display="flex"
@@ -610,10 +403,6 @@ export const SinglePageAnalyticsView = ({ item }: Props) => {
           </Box>
         </Box>
       </Box>
-
-      {showPropertiesDialog && (
-        <PropertiesDialog onClose={() => setShowPropertiesDialog(false)} />
-      )}
       {showCompareDialog && (
         <CompareDialog onClose={() => setShowCompareDialog(false)} />
       )}
